+++ cookbooks/wt_streamingcollection/attributes/default.rb
@@ -1,5 +1,5 @@
 default['wt_streamingcollection']['user']         = "nobody"
 default['wt_streamingcollection']['group']        = "nogroup"
 default['wt_streamingcollection']['download_url'] = "http://teamcity.webtrends.corp/guestAuth/repository/download/bt75/.lastSuccessful/streamingcollection-bin.tar.gz"
-default['wt_streamingcollection']['tarball']      = "streamingcollection-bin.tar.gz"
 default['wt_streamingcollection']['port']         = "8080"
+default['wt_streamingcollection']['java_opts']    = "-Xms1024m -Djava.net.preferIPv4Stack=true"