+++ cookbooks/wt_streamingcollection/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 * Unknown
 
+## 0.0.12
+* Removed 'tarball' attribute as it is included in the download_url
+* Externalized the java options 'java_opts'
+* Externalized the java options 'jmx_port' to [:wt_monitoring][:jmx_port]
+
+## 0.0.11
+* Moved monitoring attributes to wt_monitoring
+
 ## 0.0.10
 * Searches for zookeeper were made using nodes that apply the zookeeper recipe. In our environment we
 * apply a zookeeper role instead tso the search was changed to look for the role